requirements.txt
openai-whisper
yt-dlp
ffmpeg-python

